change percentage columns to decimals

+++ app/forms/steps/capital/property_form.rb
@@ -8,8 +8,8 @@ module Steps
       attribute :bedrooms, :integer
       attribute :value, :pence
       attribute :outstanding_mortgage, :pence
-      attribute :percentage_applicant_owned, :integer
-      attribute :percentage_partner_owned, :integer
+      attribute :percentage_applicant_owned, :decimal
+      attribute :percentage_partner_owned, :decimal
       attribute :is_home_address, :value_object, source: YesNoAnswer
       attribute :has_other_owners, :value_object, source: YesNoAnswer
